+++ src/models/player.NA.js
@@ -20,36 +20,12 @@ $p.newModel({
 
         destContainer.html('');
         this.displayReady();
-        console.log("HIER")
+
         this.sendUpdate( 'error', (this.media.file[0].src!=null && this.media.errorCode===7) ? 5 : this.media.errorCode );
-        return;
-    
         
-        var mouseClick = function(evt, player){
-            if (!player.getState('AWAKENING')) {
-                ref.pp.removeListener('mousedown', arguments.callee);
-                ref._setState('completed');
-            }
-        };
-        
-        this.displayReady();
-
-        if (this.pp.getConfig('enableTestcard') && !this.pp.getIsMobileClient()) {
-            this.pp.addListener('mousedown', mouseClick);
-            this._setState('error');
-            this.setTestcard( (this.media.file[0].src!=null && this.media.errorCode===7) ? 5 : this.media.errorCode);            
-        } else {
-            // this.applyImage(this.media.config.poster, destContainer);
-            this.applyCommand ('stop');
-            window.location.href = this.media.file[0].src;            
+        if (!this.pp.getConfig('enableTestcard')) {
+            window.location.href = this.media.file[0].src;      
         }
-
-    },
-    
-    detachMedia: function() {
-        this.pp.removeListener('leftclick', this.mouseClick);      
     }
-    
-
 });
 });